requirements.txt
fastapi
uvicorn
httpx
langchain
langgraph
python-dotenv
duckduckgo-search
langchain-openai
langchain-groq